Some simple macro for bulk store + recall of the memory registers

+++ common/src/main/kotlin/com/github/tmarsteel/bytecode/binary/BytecodeReader.kt
@@ -27,7 +27,7 @@ class BytecodeReader(private val input: InputStream, val closeOnEOF: Boolean = t
 
     private fun readNext(): Instruction? {
         val opcode = try {
-            readByte()
+            readByte().toByte()
         }
         catch (ex: UnexpectedEOFException) {
             if (closeOnEOF) close()
@@ -59,7 +59,7 @@ class BytecodeReader(private val input: InputStream, val closeOnEOF: Boolean = t
                  readByte().toLong()
     }
 
-    private fun readByte(): Byte {
+    private fun readByte(): Int {
         val byte = input.read()
         if (byte == -1) {
             throw UnexpectedEOFException(offsetCounter)
@@ -67,7 +67,7 @@ class BytecodeReader(private val input: InputStream, val closeOnEOF: Boolean = t
 
         offsetCounter++
 
-        return byte.toByte()
+        return byte
     }
 
     fun close() {
